% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JEGN.R
\name{JEGN.admm}
\alias{JEGN.admm}
\title{ADMM algorithm for JEGN}
\usage{
JEGN.admm(S, lambda, alpha, n = NULL, penalize.diagonal = FALSE,
  epsilon = 1e-05, maxiter = 500, rho = 0.1, rho.incr = 1.2,
  rho.max = 1e+10)
}
\arguments{
\item{S}{A matrix (\eqn{K \times G}) of list of sample covariance matrices (\eqn{p \times p}),
where K is the number of data types and G is the number of subpopulations. The (k,g)-th
element is a \eqn{p \times p} sample covariance matrix corresponding to the k-th data type
and the g-th subpopulation.}

\item{lambda}{The tuning parameter for controlling the level of sparsity of networks.}

\item{alpha}{The tuning parameter for controlling the extend of similarity among subpopulation-specific networks.}

\item{n}{The sample size. A vector with length equal to the number of subpopulations.}

\item{penalize.diagonal}{Determines whether the sparsity penalty is applied to the diagonal.}

\item{epsilon}{The tolerance parameter for convergence criteria.}

\item{maxiter}{The maximum number of iterations for the ADMM algorithm.}

\item{rho}{The penalty parameter in the ADMM algorithm.}

\item{rho.incr}{The increase step parameter for varying penalty parameter rho.}

\item{rho.max}{The maximum value of rho.}
}
\value{
a list with the following components
\item{Omega.hat}{A matrix (\eqn{K \times G}) of list of estimated precision matrices.}
\item{R.hat}{A matrix (\eqn{K \times 1}) of list of estimated common components.}
\item{M.hat}{A matrix (\eqn{K \times G}) of list of estimated subpopulation-unique components.}
}
\description{
ADMM algorithm for JEGN (Algorithm 1 in the Supplementary Information).
For details, refer to Supplementary Section S3.3.
}
\examples{
data("TCGA.BRCA")
result = JEGN(TCGA.BRCA$Sigma, 0.95, 0.4, model = "nonparanormal", weights = "equal")
}
\references{
Xiao-Fei Zhang, Le Ou-Yang, Ting Yan, Xiaohua Hu and Hong Yan (2019),
A joint graphical model for inferring gene networks across multiple subpopulations and data types,
}
\author{
Xiao-Fei Zhang  <zhangxf@mail.ccnu.edu.cn>
}
