% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JEGN.R
\name{JEGN}
\alias{JEGN}
\title{Joint gene network inference across multiple subpopulations and data types}
\usage{
JEGN(X, lambda, alpha, model = "Gaussian", weights = "equal",
  penalize.diagonal = FALSE)
}
\arguments{
\item{X}{A matrix (\eqn{K \times G}) of list of data matrices (\eqn{n_g \times p}), where K is
the number of data types and G is the number of subpopulations. The (k,g)-th element is a
\eqn{n_g \times p} data matrix corresponding to the k-th data type and the g-th subpopulation.}

\item{lambda}{The tuning parameter for controlling the level of sparsity of the estimated networks.}

\item{alpha}{The tuning parameter for controlling the extend of similarity among the estimated
subpopulation-specific networks.}

\item{model}{A character string indicating which model is used to fit the data. "Gaussian" (default)
and "nonparanormal" can be used. If model == "Gaussian", the Gaussian graphical model will be used.
If model == "nonparanormal", the nonparanormal graphical model will be used.}

\item{weights}{Determines the putative sample size of each subpopulation's data.  Allowed values:
a vector with length equal to the number of subpopulations; "equal", giving each subpopulation
 weight 1; "sample.size", giving each subpopulation weight corresponding to its sample size.}

\item{penalize.diagonal}{Determines whether the sparsity penalty is applied to the diagonal.}
}
\value{
a list with the following components
\item{Omega.hat}{A matrix (\eqn{K \times G}) of list of estimated precision matrices.}
\item{R.hat}{A matrix (\eqn{K \times 1}) of list of estimated common components.}
\item{M.hat}{A matrix (\eqn{K \times G}) of list of estimated subpopulation-unique components.}
\item{Omega.bar}{A matrix (\eqn{1 \times G}) of list of estimated gene networks.}
\item{R.bar}{A matrix of estimated common subnetworks.}
\item{M.bar}{A matrix (\eqn{1 \times G}) of list of estimated subpopulation-unique subnetworks.}
}
\description{
The complete procedure for reconstructin gene networks from gene expression data using JEGN (Algorithm 2
in the Supplementary Information). For details, refer to Supplementary Section S3.4.
}
\examples{
data("TCGA.BRCA")
result = JEGN(TCGA.BRCA$X, 0.95, 0.4, model = "nonparanormal", weights = "equal")
}
\references{
Xiao-Fei Zhang, Le Ou-Yang, Ting Yan, Xiaohua Hu and Hong Yan (2019),
A joint graphical model for inferring gene networks across multiple subpopulations and data types,
}
\author{
Xiao-Fei Zhang  <zhangxf@mail.ccnu.edu.cn>
}
