% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.BRCA}
\alias{TCGA.BRCA}
\title{TCGA breast cancer data}
\format{An object of class \code{list} of length 4.}
\usage{
TCGA.BRCA
}
\description{
The TCGA breast cancer gene expression data used in our study.
The data are obtained from the TCGA database. They are collected from two platforms:
mRNA expression (Agilent G450 microarray) and mRNA expression (RNA sequencing).
The cancers are grouped into four subtypes: Luminal A, Luminal B, HER2-enriched,
and Basal-like. The data only include expression measure of genes that overlap
with the breast cancer pathway collected from the Kyoto Encyclopedia of Genes
and Genomes database.
}
\examples{
data("TCGA.BRCA")

}
\references{
Xiao-Fei Zhang, Le Ou-Yang, Ting Yan, Xiaohua Hu and Hong Yan (2019),
A joint graphical model for inferring gene networks across multiple subpopulations and data types,
}
\author{
Xiao-Fei Zhang  <zhangxf@mail.ccnu.edu.cn>
}
\keyword{datasets}
